requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
gspread>=5.12.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
